formatted a better table, implemented links
